ci: add build workflow
This test the build system for both
Keen- and KnottyKraken.

ci: fix build call

Use -pr instead of -pr:a.
This will use the profile as the host profile(target machine)
instead of using this for both the (the one who) build and the host.
Which SHOULD trigger a mismatch error when the build compiler
cannot support the host compiler.

ci: use default cmake

+++ components/MultiNic/MultiNicXImpl.hpp
@@ -279,8 +279,7 @@ bool available(interface::FromNodeN const &, index_t aVC) {
 void push(interface::FromNodeN const &, index_t aVC, MemoryTransport &transport) {
   // Ensure src port is set correctly.
   DBG_Assert(transport[NetworkMessageTag]->src_port == N);
-  DBG_Assert(transport[NetworkMessageTag]->vc == aVC,
-             (<< "wrong VC: " << *(transport[NetworkMessageTag])));
+  DBG_Assert(static_cast<unsigned int>(transport[NetworkMessageTag]->vc) == aVC, (<< "wrong VC: " << *(transport[NetworkMessageTag])));
 
   sendQueue[(N * cfg.VChannels) + aVC].push_back(transport);
   currSendCount++;
